big changes

+++ gaharu/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, HttpResponse
+from django.shortcuts import render, HttpResponse, redirect
 from django.http import JsonResponse
 from .models import Dataset, Model
 from django.core.files.base import ContentFile
@@ -8,6 +8,7 @@ from .libraries.feature import Morfologi
 import uuid
 import numpy as np
 import cv2
+from PIL import Image
 from pprint import pprint
 from sklearn.model_selection import StratifiedKFold
 from itertools import islice, count
@@ -25,10 +26,14 @@ from django.conf import settings
 
 
 def index(request):
-    return HttpResponse('welcome')
+    if not request.user.is_authenticated:
+        return redirect("index")
+    return render(request, 'dashboard.html')
 
 
 def data_master(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     datasets = Dataset.pdobjects.all()
     kelass = [0, 1, 2, 3]
     context = {
@@ -37,8 +42,74 @@ def data_master(request):
     }
     return render(request, 'gaharu/data-master.html', context)
 
+def tambah_data_master(request):
+    return render(request, 'gaharu/tambah-data-master.html')
+
+def proses_tambah_data_master(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
+    QueryDict = request.POST
+    image64 = QueryDict.get('image')
+    kelas = QueryDict.get('kelas')
+    formatt, imgstr = image64.split(';base64,')
+    ext = formatt.split('/')[-1]
+    filename = str(uuid.uuid4())
+    fileImg = ContentFile(base64.b64decode(imgstr), name=filename + "." + ext)
+    image = cv2.imdecode(np.fromstring(fileImg.read(), np.uint8), 1)
+
+    morfologi = Morfologi(image)
+    image_gray = morfologi.gray
+    image_binary = morfologi.cleaned.astype(int)*255
+
+    image_clean = base64.b64encode(cv2.imencode('.jpg', image)[1]).decode()
+    image_gray = base64.b64encode(cv2.imencode('.jpg', image_gray)[1]).decode()
+    image_binary = base64.b64encode(cv2.imencode('.jpg', image_binary)[1]).decode()
+
+    glcm = GLCM(image)
+
+    ROUNDING = 5
+    prd      =  round(morfologi.prd(), ROUNDING)
+    plw      =  round(morfologi.plw(), ROUNDING)
+    rect     =  round(morfologi.rect(), ROUNDING)
+    nf       =  round(morfologi.narrow_factor(), ROUNDING)
+    ar       =  round(morfologi.aspect_ratio(), ROUNDING)
+    ff       =  round(morfologi.form_factor(), ROUNDING)
+
+    idm      = round(glcm.idm(), ROUNDING)
+    entropy  = round(glcm.entropy(), ROUNDING)
+    asm      = round(glcm.asm(), ROUNDING)
+    contrast = round(glcm.contrast(), ROUNDING)
+    corr     = round(glcm.korelasi(), ROUNDING)
+
+    dataset = Dataset()
+    dataset.filename = fileImg
+    dataset.form_factor = ff
+    dataset.aspect_ratio = ar
+    dataset.rect = rect
+    dataset.narrow_factor = nf
+    dataset.prd = prd
+    dataset.plw = plw
+
+    dataset.idm = idm
+    dataset.entropy = entropy
+    dataset.asm = asm
+    dataset.contrast = contrast
+    dataset.correlation = corr
+    dataset.kelas = kelas
+    dataset.save()
+
+    success = 1
+    context = {
+        'success': success,
+        'image_clean': image_clean,
+        'image_gray': image_gray,
+        'image_binary': image_binary
+    }
+    return JsonResponse(context, safe=False)
 
 def proses_input(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     QueryDict = request.POST
     image64 = QueryDict.get('image')
     kelas = QueryDict.get('kelas')
@@ -75,11 +146,15 @@ def proses_input(request):
 
 
 def data_ekstraksi(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     return HttpResponse('welcome')
     pass
 
 
 def data_anfis(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     models = Model.objects.all()
     
     context = {
@@ -89,10 +164,14 @@ def data_anfis(request):
 
 
 def tambah_model(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     return render(request, 'gaharu/tambah-model.html')
 
 
 def proses_pelatihan(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     simpan = int(request.POST.get('simpan')
                  ) if request.POST.get('simpan') else -1
     epoch = int(request.POST.get('epoch')) if request.POST.get('epoch') else 1
@@ -195,6 +274,8 @@ def proses_pelatihan(request):
 
 
 def hapus_anfis(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     model_id = int(request.POST.get('model_id'))
     model = Model.objects.get(id=model_id)
     model.filename.delete()
@@ -238,6 +319,8 @@ def lihat_anfis(request, model_id):
 
 
 def pengujian(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     models = Model.objects.all()
     context = {
         'models': models
@@ -245,6 +328,8 @@ def pengujian(request):
     return render(request, 'gaharu/pengujian.html', context)
 
 def proses_pengujian(request):
+    if not request.user.is_authenticated:
+        return redirect("index")
     model_id = int(request.POST.get('model_id'))
     image64 = request.POST.get('image')
     formatt, imgstr = image64.split(';base64,')
@@ -253,21 +338,42 @@ def proses_pengujian(request):
     fileImg = ContentFile(base64.b64decode(imgstr), name=filename + "." + ext)
     image = cv2.imdecode(np.fromstring(fileImg.read(), np.uint8), 1)
 
-    feature = Morfologi(image)
+    morfologi = Morfologi(image)
     glcm = GLCM(image)
 
+    image_gray = morfologi.gray
+    image_binary = morfologi.cleaned.astype(int)*255
+
+    image_clean = base64.b64encode(cv2.imencode('.jpg', image)[1]).decode()
+    image_gray = base64.b64encode(cv2.imencode('.jpg', image_gray)[1]).decode()
+    image_binary = base64.b64encode(cv2.imencode('.jpg', image_binary)[1]).decode()
+
+    ROUNDING = 5
+    prd      =  round(morfologi.prd(), ROUNDING)
+    plw      =  round(morfologi.plw(), ROUNDING)
+    rect     =  round(morfologi.rect(), ROUNDING)
+    nf       =  round(morfologi.narrow_factor(), ROUNDING)
+    ar       =  round(morfologi.aspect_ratio(), ROUNDING)
+    ff       =  round(morfologi.form_factor(), ROUNDING)
+
+    idm      = round(glcm.idm(), ROUNDING)
+    entropy  = round(glcm.entropy(), ROUNDING)
+    asm      = round(glcm.asm(), ROUNDING)
+    contrast = round(glcm.contrast(), ROUNDING)
+    corr     = round(glcm.korelasi(), ROUNDING)
+
     dataset = {}
-    dataset['form_factor'] = feature.form_factor()
-    dataset['aspect_ratio'] = feature.aspect_ratio()
-    dataset['rect'] = feature.rect()
-    dataset['narrow_factor'] = feature.narrow_factor()
-    dataset['prd'] = feature.prd()
-    dataset['plw'] = feature.plw()
-    dataset['idm'] = glcm.idm()
-    dataset['entropy'] = glcm.entropy()
-    dataset['asm'] = glcm.asm()
-    dataset['contrast'] = glcm.contrast()
-    dataset['correlation'] = glcm.korelasi()
+    dataset['form_factor'] = ff
+    dataset['aspect_ratio'] = ar
+    dataset['rect'] = rect
+    dataset['narrow_factor'] = nf
+    dataset['prd'] = prd
+    dataset['plw'] = plw
+    dataset['idm'] = idm
+    dataset['entropy'] = entropy
+    dataset['asm'] = asm
+    dataset['contrast'] = contrast
+    dataset['correlation'] = corr
 
     test_data = np.fromiter(dataset.values(), dtype=float).reshape(1,-1)
 
@@ -282,16 +388,29 @@ def proses_pengujian(request):
     model_anfis = model_file['model'];
     test_data_scaled = scaler.transform(test_data)
     predicted =  predict_pengujian(model_anfis, test_data_scaled)
-
-    print(predicted)
-
-    print(test_data_scaled)
+    variables = {
+        "prd": prd,
+        "plw": plw,
+        "rect": rect,
+        "nf": nf,
+        "ar": ar,
+        "ff": ff,
+        "idm": idm,
+        "entropy": entropy,
+        "asm": asm,
+        "contrast": contrast,
+        "corr": corr,
+    }
 
     response = {
         'success': 1,
         'predicted': predicted.item(),
         'fitur': json.dumps(dataset),
-        'fitur_scaled': list(test_data_scaled.reshape(-1))
+        'fitur_scaled': list(test_data_scaled.reshape(-1)),
+        'variables': variables,
+        'image_clean': image_clean,
+        'image_gray': image_gray,
+        'image_binary': image_binary
 
     }
     return JsonResponse(response, safe=False)
